Fix warning: implicit conversion changes signedness
visp/tutorial/segmentation/color/tutorial-hsv-segmentation-basic.cpp:79:89: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
   79 |   std::shared_ptr<vpDisplay> d_mask = vpDisplayFactory::createDisplay(mask, I.getWidth()+75, 0, "HSV mask");
      |                                       ~~~~~~~~~~~~~~~~                      ~~~~~~~~~~~~^~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation-basic.cpp:80:108: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
   80 |   std::shared_ptr<vpDisplay> d_I_segmented = vpDisplayFactory::createDisplay(I_segmented, 2*mask.getWidth()+80, 0, "Segmented frame");
      |                                              ~~~~~~~~~~~~~~~~                             ~~~~~~~~~~~~~~~~~^~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation-basic.cpp:81:118: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
   81 |   std::shared_ptr<vpDisplay> d_mask_hsv = vpDisplayFactory::createDisplay(maskHSV, I.getWidth()+75, mask.getHeight() + 80, "HSV mask using vpHSV");
      |                                           ~~~~~~~~~~~~~~~~                                          ~~~~~~~~~~~~~~~~~^~~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation-basic.cpp:81:96: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
   81 |   std::shared_ptr<vpDisplay> d_mask_hsv = vpDisplayFactory::createDisplay(maskHSV, I.getWidth()+75, mask.getHeight() + 80, "HSV mask using vpHSV");
      |                                           ~~~~~~~~~~~~~~~~                         ~~~~~~~~~~~~^~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation-basic.cpp:82:143: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
   82 |   std::shared_ptr<vpDisplay> d_I_segmented_hsv = vpDisplayFactory::createDisplay(I_segmented_from_HSV, 2*mask.getWidth()+80, mask.getHeight() + 80, "Segmented frame using vpHSV");
      |                                                  ~~~~~~~~~~~~~~~~                                                            ~~~~~~~~~~~~~~~~~^~~~
visp/tutorial/segmentation/color/tutorial-hsv-segmentation-basic.cpp:82:121: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
   82 |   std::shared_ptr<vpDisplay> d_I_segmented_hsv = vpDisplayFactory::createDisplay(I_segmented_from_HSV, 2*mask.getWidth()+80, mask.getHeight() + 80, "Segmented frame using vpHSV");
      |                                                  ~~~~~~~~~~~~~~~~                                      ~~~~~~~~~~~~~~~~~^~~

+++ tutorial/segmentation/color/tutorial-hsv-segmentation-basic.cpp
@@ -76,14 +76,14 @@ int main()
 #if defined(VISP_HAVE_DISPLAY)
 #if (VISP_CXX_STANDARD >= VISP_CXX_STANDARD_11)
   std::shared_ptr<vpDisplay> d_I = vpDisplayFactory::createDisplay(I);
-  std::shared_ptr<vpDisplay> d_mask = vpDisplayFactory::createDisplay(mask, I.getWidth()+75, 0, "HSV mask");
-  std::shared_ptr<vpDisplay> d_I_segmented = vpDisplayFactory::createDisplay(I_segmented, 2*mask.getWidth()+80, 0, "Segmented frame");
-  std::shared_ptr<vpDisplay> d_mask_hsv = vpDisplayFactory::createDisplay(maskHSV, I.getWidth()+75, mask.getHeight() + 80, "HSV mask using vpHSV");
-  std::shared_ptr<vpDisplay> d_I_segmented_hsv = vpDisplayFactory::createDisplay(I_segmented_from_HSV, 2*mask.getWidth()+80, mask.getHeight() + 80, "Segmented frame using vpHSV");
+  std::shared_ptr<vpDisplay> d_mask = vpDisplayFactory::createDisplay(mask, static_cast<int>(I.getWidth())+75, 0, "HSV mask");
+  std::shared_ptr<vpDisplay> d_I_segmented = vpDisplayFactory::createDisplay(I_segmented, 2*static_cast<int>(mask.getWidth())+80, 0, "Segmented frame");
+  std::shared_ptr<vpDisplay> d_mask_hsv = vpDisplayFactory::createDisplay(maskHSV, static_cast<int>(I.getWidth())+75, static_cast<int>(mask.getHeight()) + 80, "HSV mask using vpHSV");
+  std::shared_ptr<vpDisplay> d_I_segmented_hsv = vpDisplayFactory::createDisplay(I_segmented_from_HSV, 2*static_cast<int>(mask.getWidth())+80, static_cast<int>(mask.getHeight()) + 80, "Segmented frame using vpHSV");
 #else
   vpDisplay *d_I = vpDisplayFactory::allocateDisplay(I, 0, 0, "Current frame");
-  vpDisplay *d_mask = vpDisplayFactory::allocateDisplay(mask, I.getWidth()+75, 0, "HSV mask");
-  vpDisplay *d_I_segmented = vpDisplayFactory::allocateDisplay(I_segmented, 2*mask.getWidth()+80, 0, "Segmented frame");
+  vpDisplay *d_mask = vpDisplayFactory::allocateDisplay(mask, static_cast<int>(I.getWidth())+75, 0, "HSV mask");
+  vpDisplay *d_I_segmented = vpDisplayFactory::allocateDisplay(I_segmented, 2*static_cast<int>(mask.getWidth())+80, 0, "Segmented frame");
 #endif
 
   vpDisplay::display(I);
